move LineFollower part into it's own file
